requirements.txt
django-pagedown
django-allauth
django-markdownx
draceditor
django-bootstrap3
django-crispy-forms
django-braces
django-autocomplete-light